+++ internal/provider/data_source_http_test.go
@@ -1,0 +1,263 @@
+package provider
+
+import (
+	"fmt"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+)
+
+func TestDataSource_200(t *testing.T) {
+	testHttpMock := setUpMockHttpServer()
+	defer testHttpMock.server.Close()
+
+	resource.UnitTest(t, resource.TestCase{
+		ProtoV6ProviderFactories: protoV6ProviderFactories(),
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(`
+							data "http" "http_test" {
+								url = "%s/200"
+							}`, testHttpMock.server.URL),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.http.http_test", "response_body", "1.0.0"),
+					resource.TestCheckResourceAttr("data.http.http_test", "response_headers.Content-Type", "text/plain"),
+					resource.TestCheckResourceAttr("data.http.http_test", "response_headers.X-Single", "foobar"),
+					resource.TestCheckResourceAttr("data.http.http_test", "response_headers.X-Double", "1, 2"),
+					resource.TestCheckResourceAttr("data.http.http_test", "status_code", "200"),
+				),
+			},
+		},
+	})
+}
+
+func TestDataSource_404(t *testing.T) {
+	testHttpMock := setUpMockHttpServer()
+	defer testHttpMock.server.Close()
+
+	resource.UnitTest(t, resource.TestCase{
+		ProtoV6ProviderFactories: protoV6ProviderFactories(),
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(`
+							data "http" "http_test" {
+								url = "%s/404"
+							}`, testHttpMock.server.URL),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.http.http_test", "response_body", ""),
+					resource.TestCheckResourceAttr("data.http.http_test", "status_code", "404"),
+				),
+			},
+		},
+	})
+}
+
+func TestDataSource_withAuthorizationRequestHeader_200(t *testing.T) {
+	testHttpMock := setUpMockHttpServer()
+	defer testHttpMock.server.Close()
+
+	resource.UnitTest(t, resource.TestCase{
+		ProtoV6ProviderFactories: protoV6ProviderFactories(),
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(`
+							data "http" "http_test" {
+								url = "%s/restricted"
+
+								request_headers = {
+									"Authorization" = "Zm9vOmJhcg=="
+								}
+							}`, testHttpMock.server.URL),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.http.http_test", "response_body", "1.0.0"),
+					resource.TestCheckResourceAttr("data.http.http_test", "status_code", "200"),
+				),
+			},
+		},
+	})
+}
+
+func TestDataSource_withAuthorizationRequestHeader_403(t *testing.T) {
+	testHttpMock := setUpMockHttpServer()
+	defer testHttpMock.server.Close()
+
+	resource.UnitTest(t, resource.TestCase{
+		ProtoV6ProviderFactories: protoV6ProviderFactories(),
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(`
+							data "http" "http_test" {
+  								url = "%s/restricted"
+
+  								request_headers = {
+    								"Authorization" = "unauthorized"
+  								}
+							}`, testHttpMock.server.URL),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.http.http_test", "response_body", ""),
+					resource.TestCheckResourceAttr("data.http.http_test", "status_code", "403"),
+				),
+			},
+		},
+	})
+}
+
+func TestDataSource_utf8_200(t *testing.T) {
+	testHttpMock := setUpMockHttpServer()
+	defer testHttpMock.server.Close()
+
+	resource.UnitTest(t, resource.TestCase{
+		ProtoV6ProviderFactories: protoV6ProviderFactories(),
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(`
+							data "http" "http_test" {
+  								url = "%s/utf-8/200"
+							}`, testHttpMock.server.URL),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.http.http_test", "response_body", "1.0.0"),
+					resource.TestCheckResourceAttr("data.http.http_test", "response_headers.Content-Type", "text/plain; charset=UTF-8"),
+					resource.TestCheckResourceAttr("data.http.http_test", "status_code", "200"),
+				),
+			},
+		},
+	})
+}
+
+func TestDataSource_utf16_200(t *testing.T) {
+	testHttpMock := setUpMockHttpServer()
+	defer testHttpMock.server.Close()
+
+	resource.UnitTest(t, resource.TestCase{
+		ProtoV6ProviderFactories: protoV6ProviderFactories(),
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(`
+							data "http" "http_test" {
+  								url = "%s/utf-16/200"
+							}`, testHttpMock.server.URL),
+				// This should now be a warning, but unsure how to test for it...
+				// ExpectWarning: regexp.MustCompile("Content-Type is not a text type. Got: application/json; charset=UTF-16"),
+			},
+		},
+	})
+}
+
+func TestDataSource_x509cert(t *testing.T) {
+	testHttpMock := setUpMockHttpServer()
+	defer testHttpMock.server.Close()
+
+	resource.UnitTest(t, resource.TestCase{
+		ProtoV6ProviderFactories: protoV6ProviderFactories(),
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(`
+							data "http" "http_test" {
+  								url = "%s/x509-ca-cert/200"
+							}`, testHttpMock.server.URL),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.http.http_test", "response_body", "pem"),
+					resource.TestCheckResourceAttr("data.http.http_test", "status_code", "200"),
+				),
+			},
+		},
+	})
+}
+
+func TestDataSource_UpgradeFromVersion2_2_0(t *testing.T) {
+	testHttpMock := setUpMockHttpServer()
+	defer testHttpMock.server.Close()
+
+	resource.Test(t, resource.TestCase{
+		Steps: []resource.TestStep{
+			{
+				ExternalProviders: map[string]resource.ExternalProvider{
+					"http": {
+						VersionConstraint: "2.2.0",
+						Source:            "hashicorp/http",
+					},
+				},
+				Config: fmt.Sprintf(`
+							data "http" "http_test" {
+								url = "%s/200"
+							}`, testHttpMock.server.URL),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.http.http_test", "response_body", "1.0.0"),
+					resource.TestCheckResourceAttr("data.http.http_test", "response_headers.Content-Type", "text/plain"),
+					resource.TestCheckResourceAttr("data.http.http_test", "response_headers.X-Single", "foobar"),
+					resource.TestCheckResourceAttr("data.http.http_test", "response_headers.X-Double", "1, 2"),
+				),
+			},
+			{
+				ProtoV6ProviderFactories: protoV6ProviderFactories(),
+				Config: fmt.Sprintf(`
+							data "http" "http_test" {
+								url = "%s/200"
+							}`, testHttpMock.server.URL),
+				PlanOnly: true,
+			},
+			{
+				ProtoV6ProviderFactories: protoV6ProviderFactories(),
+				Config: fmt.Sprintf(`
+							data "http" "http_test" {
+								url = "%s/200"
+							}`, testHttpMock.server.URL),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.http.http_test", "response_body", "1.0.0"),
+					resource.TestCheckResourceAttr("data.http.http_test", "response_headers.Content-Type", "text/plain"),
+					resource.TestCheckResourceAttr("data.http.http_test", "response_headers.X-Single", "foobar"),
+					resource.TestCheckResourceAttr("data.http.http_test", "response_headers.X-Double", "1, 2"),
+					resource.TestCheckResourceAttr("data.http.http_test", "status_code", "200"),
+				),
+			},
+		},
+	})
+}
+
+type TestHttpMock struct {
+	server *httptest.Server
+}
+
+func setUpMockHttpServer() *TestHttpMock {
+	Server := httptest.NewServer(
+		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			w.Header().Set("Content-Type", "text/plain")
+			w.Header().Add("X-Single", "foobar")
+			w.Header().Add("X-Double", "1")
+			w.Header().Add("X-Double", "2")
+
+			switch r.URL.Path {
+			case "/200":
+				w.WriteHeader(http.StatusOK)
+				_, _ = w.Write([]byte("1.0.0"))
+			case "/restricted":
+				if r.Header.Get("Authorization") == "Zm9vOmJhcg==" {
+					w.WriteHeader(http.StatusOK)
+					_, _ = w.Write([]byte("1.0.0"))
+				} else {
+					w.WriteHeader(http.StatusForbidden)
+				}
+			case "/utf-8/200":
+				w.Header().Set("Content-Type", "text/plain; charset=UTF-8")
+				w.WriteHeader(http.StatusOK)
+				_, _ = w.Write([]byte("1.0.0"))
+			case "/utf-16/200":
+				w.Header().Set("Content-Type", "application/json; charset=UTF-16")
+				w.WriteHeader(http.StatusOK)
+				_, _ = w.Write([]byte("1.0.0"))
+			case "/x509-ca-cert/200":
+				w.Header().Set("Content-Type", "application/x-x509-ca-cert")
+				w.WriteHeader(http.StatusOK)
+				_, _ = w.Write([]byte("pem"))
+			default:
+				w.WriteHeader(http.StatusNotFound)
+			}
+		}),
+	)
+
+	return &TestHttpMock{
+		server: Server,
+	}
+}