+++ internal/provider/provider.go
@@ -1,17 +1,33 @@
 package provider
 
 import (
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-framework/diag"
+	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 )
 
-func New() *schema.Provider {
-	return &schema.Provider{
-		Schema: map[string]*schema.Schema{},
+func New() tfsdk.Provider {
+	return &provider{}
+}
 
-		DataSourcesMap: map[string]*schema.Resource{
-			"http": dataSource(),
-		},
+var _ tfsdk.Provider = (*provider)(nil)
 
-		ResourcesMap: map[string]*schema.Resource{},
-	}
+type provider struct{}
+
+func (p *provider) GetSchema(context.Context) (tfsdk.Schema, diag.Diagnostics) {
+	return tfsdk.Schema{}, nil
+}
+
+func (p *provider) Configure(context.Context, tfsdk.ConfigureProviderRequest, *tfsdk.ConfigureProviderResponse) {
+}
+
+func (p *provider) GetResources(context.Context) (map[string]tfsdk.ResourceType, diag.Diagnostics) {
+	return map[string]tfsdk.ResourceType{}, nil
+}
+
+func (p *provider) GetDataSources(context.Context) (map[string]tfsdk.DataSourceType, diag.Diagnostics) {
+	return map[string]tfsdk.DataSourceType{
+		"http": &httpDataSourceType{},
+	}, nil
 }