+++ internal/provider/provider_test.go
@@ -1,20 +1,13 @@
 package provider
 
 import (
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-framework/providerserver"
+	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
 )
 
-//nolint:unparam // error is always nil
-func testProviders() map[string]func() (*schema.Provider, error) {
-	return map[string]func() (*schema.Provider, error){
-		"http": func() (*schema.Provider, error) { return New(), nil },
-	}
-}
-
-func TestProvider(t *testing.T) {
-	if err := New().InternalValidate(); err != nil {
-		t.Fatalf("err: %s", err)
+//nolint:unparam
+func protoV6ProviderFactories() map[string]func() (tfprotov6.ProviderServer, error) {
+	return map[string]func() (tfprotov6.ProviderServer, error){
+		"http": providerserver.NewProtocol6WithError(New()),
 	}
 }